moved Images folder into components
